Fix bug of the size of an OF10 FlowRemoved message

+++ openflowjava-extension/src/main/java/org/opendaylight/netide/openflowjava/protocol/impl/serialization/factories/OF10FlowRemovedMessageFactory.java
@@ -35,11 +35,13 @@ public class OF10FlowRemovedMessageFactory implements OFSerializer<FlowRemovedMe
     @Override
     public void serialize(FlowRemovedMessage message, ByteBuf outBuffer) {
         ByteBufUtils.writeOFHeader(MESSAGE_TYPE, message, outBuffer, EncodeConstants.EMPTY_LENGTH);
+
         OFSerializer<MatchV10> matchSerializer = registry
                 .getSerializer(new MessageTypeKey<>(message.getVersion(), MatchV10.class));
+
         matchSerializer.serialize(message.getMatchV10(), outBuffer);
 
-        outBuffer.writeInt(message.getCookie().intValue());
+        outBuffer.writeLong(message.getCookie().longValue());
         outBuffer.writeShort(message.getPriority());
         outBuffer.writeByte(message.getReason().getIntValue());
         outBuffer.writeZero(PADDING);
